requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
progressbar2==4.3.2
python-dotenv==1.0.1
python-utils==3.8.2
requests==2.31.0
soupsieve==2.5
typing_extensions==4.9.0
urllib3==2.2.0
